pyproject.toml
[tool.poetry]
name = "my-mlops"
version = "0.1.0"
description = ""
authors = ["Jithish J <135303624+jithsg@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "my_mlops"}]

[tool.poetry.dependencies]
python = "^3.11"
pytorch-lightning = "^2.1.3"
wandb = "^0.16.2"
python-dotenv = "^1.0.0"
torchvision = "^0.16.2"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
hydra-colorlog = "^1.2.0"
dvc = "^3.40.1"
dvc-gdrive = "^3.0.1"
onnx = "^1.15.0"
onnxruntime = "^1.16.3"
ipynb = "^0.5.1"
ipython = "^8.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
